# todo-frontend/src/todo-api/src/models/todo-dto.ts
/* tslint:disable */
import { CategoryDto } from './category-dto';
export interface TodoDto {
  category?: CategoryDto;
  description?: string;
  done?: boolean;
  favorite?: boolean;
  id?: number;
  startDate?: string;
  title?: string;
}